# contracts/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Lpsaring API Contract
  version: 1.0.0
  description: |
    Source of truth kontrak API domain prioritas (Auth, Profile, Devices, Transactions).
    Ruang lingkup ini sengaja fokus untuk mengurangi drift backend-frontend pada flow kritikal.
    Revisi internal 2026-02-26: refactor wiring transaksi/payment tanpa perubahan signature endpoint.
servers:
  - url: /api
    description: Relative API base path
...
paths:
...
  /admin/quota-requests:
    get:
      tags: [AdminRequests]
      summary: List quota requests (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Quota request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQuotaRequestListResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /admin/quota-requests/{request_id}/process:
    post:
      tags: [AdminRequests]
      summary: Process quota request (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: request_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminQuotaRequestProcessRequest'
      responses:
        '200':
          description: Request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  responses:
...
    ErrorUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorNotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: AUTH_UNAUTHORIZED
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
        request_id:
          type: string
    ValidationErrorDetail:
      type: object
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
...
    AdminQuotaRequestItem:
      type: object
      required: [id, user_id, request_type, requested_at, status]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        request_type:
          type: string
          enum: [QUOTA, UNLIMITED]
        requested_mb:
          type: integer
          nullable: true
        requested_days:
          type: integer
          nullable: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        note:
          type: string
          nullable: true
        requested_at:
          type: string
          format: date-time
    AdminQuotaRequestListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminQuotaRequestItem'
    AdminQuotaRequestProcessRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
        approved_mb:
          type: integer
          nullable: true
        approved_days:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
...